+++ civis/ml/_model.py
@@ -185,31 +185,6 @@ def _load_estimator(job_id, run_id, filename='estimator.pkl', client=None):
     return obj
 
 
-def _exception_from_logs(exc, job_id, run_id, client, nlog=15):
-    """Create an exception if the log has a recognizable error
-
-    Search "error" emits in the last ``n_log`` lines.
-    This function presently recognizes the following errors:
-
-    - MemoryError
-    """
-    logs = client.scripts.list_containers_runs_logs(job_id, run_id, limit=nlog)
-
-    # Check for memory errors
-    msgs = [x['message'] for x in logs if x['level'] == 'error']
-    mem_err = [m for m in msgs if m.startswith('Process ran out of its')]
-    if mem_err:
-        exc = MemoryError(mem_err[0])
-    else:
-        # Unknown error; return logs to the user as a sort of traceback
-        all_logs = '\n'.join([x['message'] for x in logs])
-        if isinstance(exc, CivisJobFailure):
-            exc.error_message = all_logs + '\n' + exc.error_message
-        else:
-            exc = CivisJobFailure(all_logs)
-    return exc
-
-
 def _parse_warning(warn_str):
     """Reverse-engineer a warning string
 
@@ -476,27 +451,25 @@ class ModelFuture(ContainerFuture):
 
     def __init__(self, job_id, run_id, train_job_id=None, train_run_id=None,
                  polling_interval=None, client=None, poll_on_creation=True):
-        super().__init__(job_id, run_id,
-                         polling_interval=polling_interval,
-                         client=client,
-                         poll_on_creation=poll_on_creation)
         if train_job_id and train_run_id:
             self.is_training = False
             self.train_job_id = train_job_id
             self.train_run_id = train_run_id
         else:
             self.is_training = True
-            self.train_job_id = self.job_id
-            self.train_run_id = self.run_id
+            self.train_job_id = job_id
+            self.train_run_id = run_id
         self._metadata, self._val_metadata = None, None
         self._train_data, self._train_data_fname = None, None
         self._train_metadata = None
         self._table, self._estimator = None, None
-        self._exception_handled = False
-        self.add_done_callback(self._set_model_exception)
+        super().__init__(job_id, run_id,
+                         polling_interval=polling_interval,
+                         client=client,
+                         poll_on_creation=poll_on_creation)
 
     @staticmethod
-    def _set_model_exception(fut):
+    def _set_job_exception(fut):
         """Callback: On job completion, check the metadata.
         If it indicates an exception, replace the generic
         ``CivisJobFailure`` by a more informative ``ModelError``.
@@ -527,7 +500,7 @@ class ModelFuture(ContainerFuture):
             # If there's no metadata file
             # (we get FileNotFound or CivisJobFailure),
             # check the tail of the log for a clearer exception.
-            exc = _exception_from_logs(exc, fut.job_id, fut.run_id, fut.client)
+            exc = fut._exception_from_logs(exc)
             fut.set_exception(exc)
         except futures.CancelledError:
             # We don't need to change the exception if the run was cancelled
@@ -558,7 +531,7 @@ class ModelFuture(ContainerFuture):
         self.client = APIClient()
         self.poller = self.client.scripts.get_containers_runs
         self._begin_tracking()
-        self.add_done_callback(self._set_model_exception)
+        self.add_done_callback(self._set_job_exception)
 
     @property
     def state(self):